Fix gridAlign() in exclusive selection and clean
Implement array Sbls.travelingDefiers and use it to gridAlign() subbles to mubbles that have had other momentum than their subbles, in other words, mubbles that have not moved if their subbles moved and that have moved if their subbles did not. This is when the user has exclusive selection.
For this, use the new method Sbls.lowestSubble().
Reset travelers and travelingDefiers arrays in the end of Sbls.collisionOther().

Make moveTravelers() take the array of subbles that it acts on as a parameter.
Minor clean ups.