Move travelingDefiers logic into gridAlign
also, add a Sbls.subtree generator method, convert Sbls.travelers to a
Set, and remove travelingDefiers which is no longer needed.

Perhaps subtree is not the best name for this method since the hierarchy
is not in fact a tree, but rather a Directed Acyclic Graph, but I don't
know the corresponding term for DAGs. It's not "subgraph", that's
something else.